[REFACTOR/#162] 데이터 클래스 패키징 변경
